middleware/protect.py
from typing import Optional
from fastapi import Depends, HTTPException, status
from jose import JWTError, jwt

from helpers.auth import oauth2_scheme
from helpers.auth import SECRET_KEY
from helpers.auth import ALGORITHM

async def protect(token: str = Depends(oauth2_scheme)) -> Optional[str]:
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        email: str = payload.get("sub") # type: ignore
        if email is None:
            raise HTTPException(    
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Could not validate credentials",
                headers={"WWW-Authenticate": "Bearer"},
            )
        return email
    except JWTError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Could not validate credentials",
            headers={"WWW-Authenticate": "Bearer"},
        )